"whereami()" function: Print coordinates as integers.
The floating point part of the coordinates are distracting and not relevant.

SVN-Revision: 6568

+++ src/scripting/functions.cpp
@@ -252,7 +252,7 @@ void whereami()
 {
   if (!validate_sector_player()) return;
   ::Player* tux = Sector::current()->player;
-  log_info << "You are at x " << tux->get_pos().x << ", y " << tux->get_pos().y << std::endl;
+  log_info << "You are at x " << ((int) tux->get_pos().x) << ", y " << ((int) tux->get_pos().y) << std::endl;
 }
 
 void gotoend()
